authorisation on stream registry objects
