--- access-control/src/main/java/com/expediagroup/streamplatform/streamregistry/accesscontrol/domain/AccessControlledResource.java
@@ -1,34 +1,0 @@
-/**
- * Copyright (C) 2018-2020 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.expediagroup.streamplatform.streamregistry.accesscontrol.domain;
-
-import java.util.Map;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class AccessControlledResource {
-  private Map<String, String> keys;
-  private Map<String, String> additionalInfo;
-
-  public Map<String, String> getKeys() {
-    return keys;
-  }
-
-  public Map<String, String> getAdditionalInfo() {
-    return additionalInfo;
-  }
-}
