--- access-control/src/main/java/com/expediagroup/streamplatform/streamregistry/accesscontrol/domain/AccessType.java
@@ -1,31 +1,0 @@
-/**
- * Copyright (C) 2018-2020 Expedia, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.expediagroup.streamplatform.streamregistry.accesscontrol.domain;
-
-public enum AccessType {
-
-  CREATE("create"), UPDATE("update"), DELETE("delete");
-
-  private String accessType;
-
-  AccessType(String accessType) {
-    this.accessType = accessType;
-  }
-
-  public String getAccessType() {
-    return accessType;
-  }
-}
